fixing some bugs and cleaning up
